+++ C++/how_to_implement_segment_tree_code.cpp
